Dark Mode, Imprint, CSS
The dark mode was added: If the system scheme switches to dark, the website also adapts to the setting. The design is held in blue and saved in the new created colors.css class, in which the colors of the elements get changed.
The imprint was added: If the website goes online, it's important to have an imprint. The link refers to the thomi100.net page where the website thomas-joos.ch is added. Later, this can/might be changed.
Some smaller CSS was tweaked and the file was made better readable.