Automatically adding, added count, imprint fix
If the user enters a name in the box and clicks on "create" without "add", the name is being added automatically if it's within the criteria (e. g. accepted letters, no duplicates). Otherwise, it stays.
Furthermore, a count of the added names was introduced. It's being shown at the bottom of the page.
The imprint was shifted to the right side of the page and some smaller changes were made.